MAINT: refactor SHF to basis module
