+++ tests/utils/test_maths.py
@@ -10,7 +10,6 @@ from matgl.utils.maths import (
     SPHERICAL_BESSEL_ROOTS,
     GaussianExpansion,
     SphericalBesselFunction,
-    SphericalHarmonicsFunction,
     broadcast_states_to_atoms,
     broadcast_states_to_bonds,
     combine_sbf_shf,
@@ -21,6 +20,7 @@ from matgl.utils.maths import (
     spherical_bessel_smooth,
     unsorted_segment_fraction,
 )
+from matgl.basis._spherical import SphericalHarmonicsFunction
 
 
 class TestMath(unittest.TestCase):
